+++ include/seqan/binning_directory/binning_directory_interleaved_bloom_filter.h
@@ -256,7 +256,7 @@ public:
      * \param text Text to count occurences for.
      */
     template<typename THashCount, typename TAnyString>
-    void count(std::vector<uint64_t> & counts, TAnyString const & text) const
+    void count(std::vector<uint16_t> & counts, TAnyString const & text) const
     {
         // if (std::is_same<TValue, Dna>::value)
         //     std::cerr << "DNA\n";
@@ -327,7 +327,7 @@ public:
     }
 
     template<typename THashCount, typename TAnyString>
-    void count(std::vector<uint64_t> & counts, TAnyString const & text, uint32_t & threshold) const
+    void count(std::vector<uint16_t> & counts, TAnyString const & text, uint32_t & threshold) const
     {
         // if (std::is_same<TValue, Dna>::value)
         //     std::cerr << "DNA\n";
